adjust according to performance profiling
optimize async sound loads by starting load early
allow for 0 decals

some more details on the threading changes:

with a queued renderer, having the character rendering be threaded has a lot of contention with locks,
especially in the queue.

cl_threaded_bone_setup 1 seems to work

client leaf system was causing hitches when it was not threaded.