Remove the slowdown on Explosive Headshot
Explosive headshot is considered as a powerful upgrade, but also the bread and butter of the Sniper. It is his main upgrade that allows him to do everything he has to do. However, having the upgrade apply slowdown on top of high area damage is overkill for such a strong class.