tier1: The compiler could delete the 'memset' function call, which is used to flush 'Foo' buffer
Unfortunately, the code above may leave the buffer uncleared. Note that the 'm_count | finalcount'
arrays are cleared at the end and is not used anymore.  That's why when building the Release version
of the application, the compiler will most likely delete the call of the memset() function.

The compiler has an absolute right to do that.  This change does not affect the observed behavior
which is described in the Standard as a sequence of calls of input-output functions and volatile data
read-write functions.  That is, from the viewpoint of the C/C++ language removing the call of the
memset() function does not change anything!

See https://www.viva64.com/en/w/v597/print/, https://www.viva64.com/en/b/0388/