vgui: Remove broken mouse clipping
* Performance cost is negligible.
* When you clip cursor, you should clip to screen
coordinates of the panel window, not to relative to
the game window ones.  So initial impl clipped cursor
from 0,0 to ~window width,height which is incorrect
in windowed mode (window at the center of the screen).
Should convert relative to surface x,y;width,height to
screen coordinates which is not possible without platform
specific hacks.
* Clipping makes it harder to debug the app - cursor is
clipped to app window and IDE is not reachable.