Fixed not being able to change color of an item meter.
Typically you would be able to set the background and foreground colors of an item effect meter in the client scheme (ProgressBar.BgColor and FgColor respectively). However there's a function that is used only by a select few items, that turns the bar red once full - if the bar is not full, the set foreground color is overwritten to be white.

See: https://github.com/ValveSoftware/Source-1-Games/issues/2498