experiment: use burst mode for IO
IO threads DO get called by the main thread,
but the main thread ends up doing the IO thread jobs anyway.
how about we let the IO threads continue to
swarm over the jobs,and then let the main thread pick up
after it, doing whatever work it needs to do.
a prerequisite to every call is that the queue is processed.
so why not process it on the threads?