perf(server): compile out query cache by default
The query cache is only used in L4D, yet, still updates on all game servers.

Due to its usage of the thread pool system, and its rather complex, shared data design,
this can introduce bottlenecks per frame as the system deals with useless work/bookkeeping.

This has been profiled to use a non-negligible amount of server frame time,
and has no usages in any character system (was #ifdef guarded for TERROR)
so it is removed.