gameplay: fix switch from bonuses not considering deploy time bonuses
switch from bonus is only applied if the weapon we are switching
from has fully deployed. however, the calculation for "fully deployed"
was incorrect. it used the default switch time of 0.5, instead of also
taking into account weapon attributes which affect deploy time.

this affected the degreaser most notably, which would not
allow pyros to switch to their flare gun with increased switch speed
as soon as the degreaser deployed (0.33s).
instead, players would have to wait until 0.5s had passed before switching
which is unintuitive

ref: https://github.com/ValveSoftware/Source-1-Games/issues/3488