fix interpolation being off by 1 tick
I'm not sure about this change, but it seems to work fine
and I'm interpreting the intention to mean to interpolate
extra if the server is simulating on alternate ticks.

Predicated/client created entities do this very thing as you
can see in the TICK_INTERVAL * serverTickMultiple return path.

However, in these return paths, a server tick multiple of 1 adds on
another tick of interp when the interp is already set to a certain amount of
ticks by the user