fix: load embedded navmeshes from workshop maps
This was fixed in live tf2, but that fix broke loading navmeshes
generated using the `nav_generate` command.

The original issue was that the canonical name for a workshop map is
formatted as "workshop/pl_aceholder.ugc1234". When the game tried to
load a navmesh from the BSP, it would try to load "maps/workshop/pl_aceholder.ugc1234.nav",
instead of "maps/pl_aceholder.nav". The fix currently implemented in
live correctly uses the "clean" name for the BSP, but also uses the
clean name when looking in the MOD directories (effectively everything
*except* the BSP). Since `nav_generate` saves its navmeshes with the
canonical name, they would never be found.