perf(load): lazily load parsed data for game DLL when on a listen server
during startup, the particles manifest and item schema were being parsed
in both the server DLL and client DLL on startup. this greatly delays
a player's ability quickly launch the game and queue into a match.

so, we can instead just skip initializing for server DLL until we need the manifests
at map load. particles are already parsed specifically for the map at level init,
and pending item schema updates from the GC are also initialized at level init,
so this should be a fine change. it also shouldn't touch the dedicated server path.

as an added boost, we now also backport a change from CSGO which added
threaded init to the client. this completely hides the particle init cost
behind item schema parsing time, which is the longer of the two initializations.