perf: shared object find optimizations
when an econ item view goes looking for a
econ item, it looks through the CSharedObjectTypeCache
ideally, this would maintain a map instead of looking up
with O(n) search, but that's a more complicated and scary
change

instead, skip the amount of virtual GetTypeID calls since
SharedObject::BIsKeyEqual is only used in CSharedObjectTypeCache
so the equal types are guaranteed. on GC, BIsKeyEqual is used in
CSharedObjectTransaction as well, so we keep the old check for it.

also move to a static_cast for CEconItem to be explicit/efficient with casting