perf: fix FPS drops with mouse movement (esp. on high rate)
on high rate mice (8KHz like the Razer Viper),
the message queue can get extremely long

we replicate an SDL workaround for this:

https://github.com/libsdl-org/SDL/commit/afb9ff9507f013cc413f091f558479433508ba83

this fixes a performance drop from 1000FPS to 40FPS while moving such a mouse quickly

now, the FPS drops only slightly.

note that there was a subsequent patch in SDL which reduces the
effectiveness of this fix, I have not applied it here because it
was a speculative fix for an issue that was not reproducible in TF2

github.com/ libsdl-org/SDL/pull/4831

"I believe this was part of the cause of the unexpected warping"

with that patch applied, the FPS drop is from 1000FPS to 120FPS
instead of 1000FPS to 950FPS. still an improvement but to slight to
really bother.