gameplay: make caber damage consistent
since the caber used distance mod, its base damage
was extremely inconsistent and seemingly random,
with a player essentially needing to "aim" this
melee weapons using their eye position.

also set the explosion location to the player's center, so that
the explosion falloff is much more consistent and easier to
land optimally.

we set DMG_NOCLOSEDISTANCEMOD because there are various parts of
the code which adjust the damage when this flag is not present,
without checking if DMG_USEDISTANCEMOD is really applied.

also treat the caber as a grenade in all cases in the code,
for consistency.

finally, add missing charge mini-crit code for the caber and
pain train.

ref: https://github.com/ValveSoftware/Source-1-Games/issues/4428