gameplay: make buffalo steak sandvich not reduce speed
the buffalo steak sandvich was made to not stack with other
movement speed bonuses. however, the way this was implemented
nullifies the greater of the two movement speed.

this fix makes it so if we are already above the berzerker speed cap,
we just don't apply berzerker speed at all and retain the faster speed.

if we are below the cap, then we boost by 35% speed like normal, up
to the berzerker cap of 135%

this only applies to MvM movement speed bonuses, since nothing
else can exceed the 135% cap in the base game.

ref: https://github.com/ValveSoftware/Source-1-Games/issues/3476