gameplay: consider overheal penalties when calculating pct hp charge rate
this was fixed for the Razorback, but the Fists of Steel can be
used to build uber just as fast since even at max overheal, the Heavy
will not hit the reduced charge rate health amount.

I'm not sure how the Fists of Steel was implemented, but I tried to
provide an example of how this bug could be fixed.