fix: round networking rates to nearest tick interval on boundaries
This fixes an alignment issue with tickrate vs. networking rate.
Now, if our networking rate is within 1 unit of a tick interval,
we just jump to that tick interval. This makes for slightly
faster/slower networking packet rates inaccurate to the set
integer rate, but the difference is miniscule and provides
many benefits for synchronization purposes.

In CS:GO, the tickrate was changed to a value expressable by
an integer (64), this is the next best thing.