fix: data exploit with old LCD keyboard system
The LCD system supports the configurable Logitech G-15 keyboard.

This is a very old keyboard, which is still supported by Logitech, and
has an integration in the Source Engine.

TF2 never shipped the g15.dll for full support, but there is evidence of
customers enabling it by copying it from other games.

https://www.tf2newbs.com/newbs_blog/2010/01/14/tf2_g-series

So, due to a potential oversight, despite most of it being protected
by checking for a connection to the keyboard, the g15_dumpplayer
console command was never protected.

Thus, this command has gained attention from players, even seeing some
use in community vote kick external software that hooks into the game
using commands and console logging.

However, at the same time, some players use it to gather information
because it shares the player resource, including enemy health.

With all of that context, there are a few options I see for this:

1) Delete the G15 system. It was never officialy supported or shipped
in TF2. There are legitimate users who enjoy this feature, so might
not be the best route.

2) Prevent the command from being used by people without the keyboard.
This still lets the keyboard users technically cheat with the enemy
health dump, and it also breaks vote kick software, and also
could be sensitive to DLL injection with g15.dll.

3) Remove g_PR dumping from the command. This would still allow for
most reasonable uses of information on the LCD display, for your
weapon and local player. This might be okay, but this breaks
compatibility, and also removes the potential to use it in vote
kick software.

4) Filter out the enemy health from the dump. This is the most complex
approach but avoids the impact to legitimate uses.

I've included options 2 and 3 in this patch.