fix: lack of replication for env_tonemap_controller rate
ref: https://github.com/ValveSoftware/Source-1-Games/issues/3501

This just replicates the cvar so that any
env_tonemap_controller can set the tonemap rate

This is unlike L4D/CSGO, where there is a concept of
a "master" tonemap controller which level authors
can mark. However, I don't think this is needed for TF.

Side note, because of that, there isn't a game system
available for tonemapping, like fog system, so I just
reset the value on level init in gameinterface.
Non ideal, but good enough?